// src/models/response/PricatsResponse.ts
export interface PricatsResponse{
    documentId: number;
    dateTime: string;
    dateTimeInsert: string;
    dateTimeUpdate: string;
    documentDate: string;
    edi: string;
    documentType: string;
    documentStatus: string;
    documentNumber: string;
    senderId: number;
    receiverId: number;
}